requirements.txt
numpy<2
pandas
matplotlib
seaborn
scikit-learn
opencv-python
pillow
tqdm
streamlit
pyyaml